Finished DialogBar
In this patch I've finished DialogBar, and now we can realize dialogs between two and more persons. However we can create a monolog, but to my mind it would be better to do it with notices above a hero.
P.S. Just remembered. Create a blackout of background and stop the time.